bblocks/requirements.txt
shapely
requests
pyyaml
